Add ability to get first 100 contacts with emails
Because it is not possible to know ahead-of-time how many contacts
will have emails, it's necessary to call the API server several times,
checking after each time whether enough contacts have been retrieved.
The number of contacts to retrieve each time is a balance between
retrieving too many left-over contacts which will be thrown away (high limit),
and number of requests (low limit).

Currently, the username and password are hard-coded.
The next step will be the extract these out.